+++ src/application/usecases/user/delete-user.usecase.ts
@@ -1,4 +1,4 @@
-import { ForbiddenException, Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { UserRepository } from 'src/domain/repositories/user.repository';
 
 @Injectable()
@@ -7,6 +7,7 @@ export class DeleteUserUsecase {
 
   async execute(id: string, churchId: string): Promise<void> {
     const user = await this.userRepository.findById(id);
+    if (!user) throw new NotFoundException('User not found');
     if (user.churchId !== churchId) throw new ForbiddenException();
     return this.userRepository.delete(id);
   }
